requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-dotenv
python-jose
passlib[bcrypt]